Add files via upload
<iostream> (for input and output operations)
In C++, the std (standard) namespace contains all the built-in functions and classes of the C++ Standard Library, such as cout,cin.
In C++, the main() function is the entry point of any program. It is where execution begins.
In C++, cin and cout are part of the Standard Input/Output (I/O) Stream and are defined in the <iostream> library. They allow us to take input from the user and display output on the screen.
In C++, cin and cout are part of the Standard Input/Output (I/O) Stream and are defined in the <iostream> library. They allow us to take input from the user and display output on the screen.
Use cin to take input from the user.